faet: porter modal 구현, mypage 제거 및 가이드 페이지 생성
