+++ src/components/GuideLines.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const GuideLines = () => {
+  return (
+    <div>
+      <div className='guide-line'>GuideLines</div>
+      <div className='description'>준비중...</div>
+    </div>
+  );
+};
+
+export default GuideLines;
